# Improve DateTime{Offset} "r" and "o" formatting performance