Two main changes:
1. Rewrote the formatting to use span directly rather than going through StringBuilder, only to then discover that we already had almost exactly the same implementation in Utf8Formatter.  As that one had some extra optimizations around JIT behaviors, I ported that over instead.  But I also took advantage of a few things that can be done only in corelib.
2. Avoided [ThreadStatic] lookups unless necessary.

**Before:**

 |     Method |     Mean |     Error |    StdDev |
 |----------- |---------:|----------:|----------:|
 |    ToStringO | 2.201 ms | 0.0434 ms | 0.0445 ms |
 |    ToStringR | 1.881 ms | 0.0363 ms | 0.0484 ms |
 | TryFormatO | 2.098 ms | 0.0387 ms | 0.0362 ms |
 | TryFormatR | 1.809 ms | 0.0359 ms | 0.0399 ms |

**After:**

 |     Method |     Mean |    Error |   StdDev |
 |----------- |---------:|---------:|---------:|
 |    ToStringO | 942.6 us | 5.217 us | 4.356 us |
 |    ToStringR | 591.4 us | 5.026 us | 4.701 us |
 | TryFormatO | 827.9 us | 6.094 us | 5.402 us |
 | TryFormatR | 521.1 us | 3.701 us | 3.462 us |

Test:
```C#
using System;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

[InProcess]
public class Program
{
    public static void Main() => BenchmarkRunner.Run<Program>();
    private const int Iters = 10_000;
    private static readonly DateTime s_now = DateTime.UtcNow;
    private static readonly char[] s_buffer = new char[100];

    [Benchmark]
    public void ToStringO()
    {
        DateTime dt = s_now;
        for (int i = 0; i < Iters; i++) dt.ToString("O");
    }

    [Benchmark]
    public void ToStringR()
    {
        DateTime dt = s_now;
        for (int i = 0; i < Iters; i++) dt.ToString("R");
    }

    [Benchmark]
    public void TryFormatO()
    {
        DateTime dt = s_now;
        Span<char> arr = s_buffer;
        for (int i = 0; i < Iters; i++) dt.TryFormat(arr, out int charsWritten, "O");
    }

    [Benchmark]
    public void TryFormatR()
    {
        DateTime dt = s_now;
        Span<char> arr = s_buffer;
        for (int i = 0; i < Iters; i++) dt.TryFormat(arr, out int charsWritten, "R");
    }
}
```

cc: @danmosemsft, @jkotas, @ahsonkhan